fixing issues with ControlledActionButton and renaming its file name
