+++ app/components/assets/export-button.tsx
@@ -1,6 +1,6 @@
 import { useLoaderData } from "@remix-run/react";
 import type { loader } from "~/routes/_layout+/assets._index";
-import { ControlledActionButton } from "../subscription/premium-feature-button";
+import { ControlledActionButton } from "../subscription/controlled-action-button";
 
 export const ExportButton = ({
   canExportAssets,
